requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
chardet==4.0.0
cryptoaddress==0.2.1
discord==1.7.3
discord.py==1.7.3
idna==3.2
multidict==5.2.0
numpy==1.21.2
pandas==1.3.3
pycodestyle==2.7.0
pysha3==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
six==1.16.0
tinydb==4.5.2
toml==0.10.2
typing-extensions==3.10.0.2
yarl==1.6.3
